Backport GEOT-6009 - make WMTS tests online (#2062)
* Add Online tests and less chatter
